Move autotag to internal
