+++ internal/autotag/gallery.go
