+++ internal/autotag/gallery_test.go
