+++ internal/autotag/image.go
