+++ internal/autotag/image_test.go
