+++ internal/autotag/scene.go
