+++ internal/autotag/scene_test.go
