+++ internal/autotag/studio_test.go
